requirements.txt
streamlit
tensorflow
numpy
opencv-python
pillow
matplotlib
scikit-learn
geocoder
requests
groq
reportlab